Add hasattr check to ensure that properties can be created for __geo_interface__
When Shapy creates a new geometry properties are not passed from parent objects.  An example is the intersections of two polygons creates a new polygon.  When attempting to call the Shapy view method on the intersection results pydraw uses __geo_interface__  to get the data for the polygon and raised an error due to missing properties.